+++ backend/src/main/java/team/teamby/teambyteam/feed/application/ImageUploadService.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import team.teamby.teambyteam.aws.s3.FileCloudUploader;
+import team.teamby.teambyteam.filesystem.awss3.FileCloudUploader;
 import team.teamby.teambyteam.feed.application.dto.ImageUrlResponse;
 import team.teamby.teambyteam.feed.application.dto.ImageUrlsResponse;
 import team.teamby.teambyteam.feed.application.dto.UploadImageRequest;
