+++ backend/src/main/java/team/teamby/teambyteam/filesystem/S3Uploader.java
@@ -1,4 +1,4 @@
-package team.teamby.teambyteam.aws.s3;
+package team.teamby.teambyteam.filesystem;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
@@ -6,6 +6,7 @@ import org.springframework.web.multipart.MultipartFile;
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
+import team.teamby.teambyteam.filesystem.awss3.FileCloudUploader;
 
 import java.io.IOException;
