+++ backend/src/main/java/team/teamby/teambyteam/filesystem/awss3/S3Configuration.java
@@ -1,12 +1,12 @@
-package team.teamby.teambyteam.aws.s3.configuration;
+package team.teamby.teambyteam.filesystem.awss3.configuration;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.s3.S3Client;
-import team.teamby.teambyteam.aws.s3.FileCloudUploader;
-import team.teamby.teambyteam.aws.s3.S3Uploader;
+import team.teamby.teambyteam.filesystem.awss3.FileCloudUploader;
+import team.teamby.teambyteam.filesystem.S3Uploader;
 
 @Configuration
 public class S3Configuration {
